All user deletions added to the user route

+++ backend/routes/users.js
@@ -3,7 +3,7 @@ const mongoose = require("mongoose");
 const saveImage = require("../utils/saveImage");
 const fs = require("fs");
 
-const { User, validateUser } = require("../models/user");
+const { User, validateUser, deleteUser } = require("../models/user");
 const router = express.Router();
 
 // get all the users
@@ -11,6 +11,7 @@ router.get("/", async (req, res) => {
   const users = await User.find({}).sort("name");
   res.send(users);
 });
+
 // get user by id
 router.get("/:id", async (req, res) => {
   const id = req.params.id;
@@ -20,6 +21,7 @@ router.get("/:id", async (req, res) => {
   if (user) return res.send(user);
   res.status(404).send("User not found");
 });
+
 // update user by id
 router.put("/:id", async (req, res) => {
   const id = req.params.id;
@@ -34,8 +36,9 @@ router.put("/:id", async (req, res) => {
 
   res.status(404).send("User not found");
 });
+
 // update user image
-router.patch("/:id", async (req, res) => {
+router.patch("/picture/:id", async (req, res) => {
   const picture = req.files ? req.files.picture || undefined : undefined;
   if (!picture || !/image/.test(picture.mimetype))
     return res.status(400).send("Picture required or Invalid picture type");
@@ -46,11 +49,33 @@ router.patch("/:id", async (req, res) => {
 
   const user = await User.findById(id);
   if (!user) return res.status(404).send("User not found");
-  fs.unlink(`./public/${user.picture}`, (err) => {
-    if (err) winston.error(err);
-  });
-  user.picture = await saveImage(picture, user.picture.split(".")[0]);
+
+  if (!/default.png/.test(user.picture))
+    fs.unlink(`./public/${user.picture}`, (err) => {
+      if (err && err.code !== "ENOENT") res.status(500).send(err);
+    });
+
+  user.picture = await saveImage(picture, `profilePic/${user._id}`);
+
   await user.save();
   res.send(user);
 });
+
+// delete user and its relations
+router.delete("/:id", async (req, res) => {
+  const id = req.params.id;
+  if (!mongoose.isValidObjectId(id))
+    return res.status(400).send("Invalid school id");
+
+  const user = await User.findByIdAndDelete(id);
+  if (!user) return res.status(404).send("User already deleted");
+  if (!/default.png/.test(user.picture))
+    fs.unlink(`./public/${user.picture}`, (err) => {
+      if (err && err.code !== "ENOENT") res.status(500).send(err);
+    });
+  const result = await deleteUser(id, user.role);
+  result.user = user;
+  res.send(result);
+});
+
 module.exports = router;
